+++ src/test/kotlin/com/jetbrains/snakecharm/SnakemakeTestCase.kt
@@ -23,18 +23,13 @@ import javax.swing.SwingUtilities
 abstract class SnakemakeTestCase : UsefulTestCase() {
     // TODO: could be extend SnakemakeTestCase here?
 
-    companion object {
-        const val PYTHON_2_MOCK_SDK = "2.7"
-        const val PYTHON_3_MOCK_SDK = "3.7"
-    }
-
     protected val ourPyDescriptor = PyLightProjectDescriptor(
-            PYTHON_2_MOCK_SDK,
+            LanguageLevel.PYTHON27,
             SnakemakeTestUtil.getTestDataPath().toString()
     )
 
     protected val ourPy3Descriptor = PyLightProjectDescriptor(
-            PYTHON_3_MOCK_SDK,
+            LanguageLevel.PYTHON37,
             SnakemakeTestUtil.getTestDataPath().toString(),
             SnakemakeTestUtil.getTestDataPath().resolve("MockPackages3")
     )